PWGGA/GammaConv: fix for e29444a
tested successfully on mac os x and linux ubuntu (AddTask_GammaConvV1_PbPb.C)
