+++ PWGGA/GammaConvBase/PWGGAGammaConvBaseLinkDef.h
@@ -42,4 +42,6 @@
 #pragma link C++ class AliDalitzData+;
 #pragma link C++ class AliCaloSigmaCuts+;
 
+#pragma link C++ class utils_TH1+ ;
+
 #endif
