+++ PWGGA/PWGGAUtils/PWGGAUtilsLinkDef.h
@@ -5,7 +5,5 @@
 #pragma link off all functions;
 
 #pragma link C++ class AliEPFlattener+ ;
-#pragma link C++ class TH1_ExponentialInterpolation+ ;
-#pragma link C++ class utils_TH1+ ;
 
 #endif
